support cursor expression

+++ src/include/knl/knl_session.h
@@ -554,7 +554,7 @@ typedef struct knl_u_utils_context {
     int GUCNestLevel; /* 1 when in main transaction */
 
     unsigned int b_format_behavior_compat_flags;
-    unsigned int behavior_compat_flags;
+    int64 behavior_compat_flags;
     unsigned int plsql_compile_behavior_compat_flags;
 
     int save_argc;
